pyproject.toml
[project]
name = "easyllms"
version = "0.1.0"
description = "A unified interface for streaming responses from multiple AI providers"
authors = [
    {name = "Muhammad Usman", email = "usman@revolveai.com"},
    {name = "Salman Ahmad", email = "salman.ahmad@revolveai.com"}
    ]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "openai",
    "anthropic",
    "requests",
    "python-dotenv"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["easyllms", "easyllms.providers"]